+++ cmd/mysqldef/mysqldef_test.go
@@ -225,6 +225,46 @@ func TestMysqldefCreateTableKey(t *testing.T) {
 	))
 }
 
+func TestMysqldefCreateTableForeignKey(t *testing.T) {
+	resetTestDatabase()
+
+	createUsers := "CREATE TABLE users (id BIGINT PRIMARY KEY);\n"
+	createPosts := stripHeredoc(`
+			CREATE TABLE posts (
+			  content text,
+			  user_id bigint
+			);
+			`,
+	)
+	assertApplyOutput(t, createUsers+createPosts, applyPrefix+createUsers+createPosts)
+	assertApplyOutput(t, createUsers+createPosts, nothingModified)
+
+	createPosts = stripHeredoc(`
+			CREATE TABLE posts (
+			  content text,
+			  user_id bigint,
+			  CONSTRAINT posts_ibfk_1 FOREIGN KEY (user_id) REFERENCES users (id)
+			);
+			`,
+	)
+	assertApplyOutput(t, createUsers+createPosts, applyPrefix+"ALTER TABLE posts ADD CONSTRAINT posts_ibfk_1 FOREIGN KEY (user_id) REFERENCES users (id);\n")
+	assertApplyOutput(t, createUsers+createPosts, nothingModified)
+
+	createPosts = stripHeredoc(`
+			CREATE TABLE posts (
+			  content text,
+			  user_id bigint
+			);
+			`,
+	)
+	assertApplyOutput(t, createUsers+createPosts, applyPrefix+stripHeredoc(`
+		ALTER TABLE posts DROP FOREIGN KEY posts_ibfk_1;
+		ALTER TABLE posts DROP INDEX posts_ibfk_1;
+		`,
+	))
+	assertApplyOutput(t, createUsers+createPosts, nothingModified)
+}
+
 func TestMysqldefCreateTableSyntaxError(t *testing.T) {
 	assertApplyFailure(t, "CREATE TABLE users (id bigint,);", `found syntax error when parsing DDL "CREATE TABLE users (id bigint,)": syntax error at position 32`+"\n")
 }
@@ -258,7 +298,7 @@ func TestMysqldefColumnLiteral(t *testing.T) {
 	assertApplyOutput(t, createTable, nothingModified)
 }
 
-func TestMysqldefIntegerBug(t *testing.T) {
+func TestMysqldefTypeAliases(t *testing.T) {
 	resetTestDatabase()
 
 	createTable := stripHeredoc(`